+++ plugins/builtin/apm/nomad/plugin/plugin.go
@@ -7,6 +7,7 @@ import (
 	nomadHelper "github.com/hashicorp/nomad-autoscaler/helper/nomad"
 	"github.com/hashicorp/nomad-autoscaler/plugins"
 	"github.com/hashicorp/nomad-autoscaler/plugins/apm"
+	"github.com/hashicorp/nomad-autoscaler/plugins/base"
 	"github.com/hashicorp/nomad/api"
 )
 
@@ -25,7 +26,7 @@ var (
 		Factory: func(l hclog.Logger) interface{} { return NewNomadPlugin(l) },
 	}
 
-	pluginInfo = &plugins.PluginInfo{
+	pluginInfo = &base.PluginInfo{
 		Name:       pluginName,
 		PluginType: plugins.PluginTypeAPM,
 	}
@@ -55,6 +56,6 @@ func (a *APMPlugin) SetConfig(config map[string]string) error {
 	return nil
 }
 
-func (a *APMPlugin) PluginInfo() (*plugins.PluginInfo, error) {
+func (a *APMPlugin) PluginInfo() (*base.PluginInfo, error) {
 	return pluginInfo, nil
 }