+++ plugins/builtin/strategy/target-value/main.go
@@ -1,17 +1,16 @@
 package main
 
 import (
-	"github.com/hashicorp/go-plugin"
+	hclog "github.com/hashicorp/go-hclog"
 	"github.com/hashicorp/nomad-autoscaler/plugins"
 	targetvalue "github.com/hashicorp/nomad-autoscaler/plugins/builtin/strategy/target-value/plugin"
-	"github.com/hashicorp/nomad-autoscaler/plugins/strategy"
 )
 
 func main() {
-	plugin.Serve(&plugin.ServeConfig{
-		HandshakeConfig: plugins.Handshake,
-		Plugins: map[string]plugin.Plugin{
-			plugins.PluginTypeStrategy: &strategy.Plugin{Impl: &targetvalue.StrategyPlugin{}},
-		},
-	})
+	plugins.Serve(factory)
+}
+
+// factory returns a new instance of the TargetValue Strategy plugin.
+func factory(log hclog.Logger) interface{} {
+	return targetvalue.NewTargetValuePlugin(log)
 }