+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-## Unreleased
+## UNRELEASED
+
+## 0.0.1-techpreview2 (April 9, 2020)
+
+IMPROVEMENTS:
+ * core: update Nomad API dependency to 0.11.0 [[GH-85](https://github.com/hashicorp/nomad-autoscaler/pull/85)]
+ * plugin: return an error when Prometheus APM query returns NaN [[GH-87](https://github.com/hashicorp/nomad-autoscaler/pull/87)]
+ * plugin: improve user experience when developing external plugins [[GH-82](https://github.com/hashicorp/nomad-autoscaler/pull/82)]
 
 BUG FIXES:
  * plugin: allow the internal `target-value` plugin to handle scaling to zero as well as use target values of zero [[GH-77](https://github.com/hashicorp/nomad-autoscaler/pull/77)]